backend/requirements/base.txt
Django==2.2.27
django-model-utils==3.1.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
PyJWT==1.6.4
django-cors-headers==3.0.0
django-filter==2.4.0
raven==6.0.0
psycopg2==2.8.4
sentry-sdk==1.5.4
rest-condition==1.0.3
xlrd==1.2.0
Pillow==9.1.0
drf-renderer-xlsx==0.4.1
requests==2.22.0
WeasyPrint==60.1